requirements.txt
mcdm==1.4
mrmr-selection==0.2.8
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.1.1
scanpy==1.9.6
gap-stat==2.0.3
matplotlib==3.8.0
imbalanced-learn==0.11.0
kneed==0.8.5
tables==3.8.0
leidenalg==0.10.2
scipy==1.11.4
torch

